requirements.txt
Django==5.1.6
djangorestframework
django-currentuser
factory_boy
Pillow==10.0.1
python-decouple==3.8
django-cors-headers==4.3.1
bcrypt
djangorestframework-simplejwt
lxml==5.2.2
pytest-django
pytest-cov
django-filter
drf-nested-routers
django-extensions
python-dotenv